requirements.txt
beautifulsoup4
chromadb
dashscope
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langgraph
pydantic
python-dotenv
sentence-transformers
tiktoken
typing-extensions
